+++ axum/src/api/files/file.rs
@@ -1,6 +1,6 @@
 use crate::api::error::APIError;
 use crate::auth::jwt::Claims;
-use crate::db::crud::file;
+use crate::db::crud::files::file;
 use crate::models::files::{File, FileListParams, FileUpdate};
 use crate::s3::post::delete_s3;
 use crate::schema::AppState;
