+++ axum/src/db/crud/files/mod.rs
@@ -1,0 +1,1 @@
+pub mod file;
