+++ axum/src/db/crud/user/mod.rs
@@ -1,0 +1,1 @@
+pub mod student;
